Bug 1438044 - re-enable WebVR for macOS in release
For FF59, we disabled WebVR for macOS before allowing it to ride the trains to release.  Softvision was unable to verify for QA due to challenges getting a working hardware configuration for macOS VR at SoftVision.

We have since gained approval from the Firefox Release Team to re-enable WebVR for macOS in release for FF60.

Essentially, we need to reverse the changes in bug 1426500 and uplift to FF59/Beta before the next cycle.